FDG-10796 remove generated reports from download table row
