+++ src/components/published-reports/download-report-table/download-report-table-row/download-report-table-row.spec.js
@@ -89,36 +89,4 @@ describe('Download report table row component', () => {
     expect(getByText('Download')).toBeInTheDocument();
     expect(getByRole('img', { hidden: true, name: '' })).toHaveClass('fa-cloud-arrow-down');
   });
-
-  describe('Generated report table row', () => {
-    const mockGeneratedReport = {
-      name: 'Name',
-      downloadName: 'Download Name',
-      date: '5/1/2021',
-      size: '5K',
-      config: {},
-      data: [],
-      colConfig: {},
-    };
-
-    it('renders a file row', async () => {
-      const setApiError = jest.fn();
-      const setIsLoading = jest.fn();
-      const { findByTestId } = render(
-        <DownloadReportTableRow generatedReport={mockGeneratedReport} setApiErrorMessage={setApiError} setIsLoading={setIsLoading} />
-      );
-      expect(await findByTestId('file-download-row')).toBeInTheDocument();
-    });
-
-    it('renders a download link', async () => {
-      const setApiError = jest.fn();
-      const setIsLoading = jest.fn();
-      const { findByRole } = render(
-        <DownloadReportTableRow generatedReport={mockGeneratedReport} setApiErrorMessage={setApiError} setIsLoading={setIsLoading} />
-      );
-      const downloadLink = await findByRole('link');
-      expect(downloadLink).toBeInTheDocument();
-      expect(downloadLink).toHaveAttribute('download', 'Download Name');
-    });
-  });
 });
