+++ src/components/published-reports/download-report-table/gen-report-download-button/gen-report-download-button.module.scss
@@ -1,0 +1,3 @@
+.downloadItem {
+  min-height: 2.75rem;
+}
